feat(ngx-diff): add `ngx-unified-diff` component to replace `inline-diff`; deprecate `inline-diff`
