+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ngx-diff",
-  "version": "7.0.0",
+  "version": "8.0.0-alpha.0",
   "type": "module",
   "scripts": {
     "ng": "ng",
